+++ src/common/i18n/en.js
@@ -1,0 +1,17 @@
+/*
+ * pwix:date/src/common/i18n/en.js
+ */
+
+DateJs.i18n = {
+    ...DateJs.i18n,
+    ...{
+        en: {
+            ms_to_human: {
+                day_abbr: 'day(s)',
+                hour_abbr: 'h.',
+                minute_abbr: 'min.',
+                second_abbr: 'sec.'
+            }
+        }
+    }
+};