+++ src/common/i18n/fr.js
@@ -1,0 +1,17 @@
+/*
+ * pwix:i18n/src/common/i18n/fr.js
+ */
+
+DateJs.i18n = {
+    ...DateJs.i18n,
+    ...{
+        fr: {
+            ms_to_human: {
+                day_abbr: 'jour(s)',
+                hour_abbr: 'h.',
+                minute_abbr: 'min.',
+                second_abbr: 'sec.'
+            }
+        }
+    }
+};