+++ src/common/js/functions.js
@@ -4,8 +4,10 @@
 
 import _ from 'lodash';
 import strftime from 'strftime';
+import printf from 'printf';
 
 import { Logger } from 'meteor/pwix:logger';
+import { pwixI18n } from 'meteor/pwix:i18n';
 
 const logger = Logger.get();
 
@@ -98,6 +100,27 @@ _.merge( DateJs, {
             d = new Date( date );
         }
         return Boolean( d ? !isNaN( d.getTime()) : false );
+    },
+
+    /**
+     * @summary Try to convert a ms value to a human-readable string
+     * @param {Number} value
+     * @returns {String} a human-readable string
+     */
+    msToHuman( value ){
+        const min = 60*1000;
+        const hour = 60*min;
+        const day = 24*hour;
+        if( value < min ){
+            return printf( '%.1f %s', value / 1000, pwixI18n.label( I18N, 'ms_to_human.second_abbr' ));
+        }
+        if( value < hour ){
+            return printf( '%.1f %s', value / min, pwixI18n.label( I18N, 'ms_to_human.minute_abbr' ));
+        }
+        if( value < day ){
+            return printf( '%.1f %s', value / hour, pwixI18n.label( I18N, 'ms_to_human.hour_abbr' ));
+        }
+        return printf( '%.1f %s', value / day, pwixI18n.label( I18N, 'ms_to_human.day_abbr' ));
     },
 
     /**