+++ src/common/js/index.js
@@ -5,5 +5,6 @@
 import './global.js';
 import './constants.js';
 import './configure.js';
+import './i18n.js';
 //
 import './functions.js';